gpt-query-database/requirements.txt
langchain==0.2.6
SQLAlchemy==2.0.31
pyodbc==5.1.0
openai==1.35.10
python-dotenv==1.0.1
